Move integer scaling handling into `render.cpp` from `sdlmain.cpp`
